UI updation and Gemini and OpenCage integrated
In this PR, I have somewhat changed the UI/UX of the web app.
Also, I have successfully integrated Gemini-1.5 pro latest and I'm now getting responses from Gemini.
From the story generated, I am extracting the location's name and then I'm using OpenCage API to extract the longitude and the latitude of this location.
I have also written a function to generate the KML File which will be passed on to the LG setup.